Extract SPI exclusive device into shared crate
